chore: organize stack package
Move validators to a new file, validate.go.
Cluster error messages by category.
