+++ internal/pkg/deploy/cloudformation/stack/validate.go
@@ -1,0 +1,163 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package stack
+
+import (
+	"fmt"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/copilot-cli/internal/pkg/manifest"
+)
+
+// Validate that paths contain only an approved set of characters to guard against command injection.
+// We can accept 0-9A-Za-z-_.
+func validatePath(input string, maxLength int) error {
+	if len(input) > maxLength {
+		return fmt.Errorf("path must be less than %d bytes in length", maxLength)
+	}
+	if len(input) == 0 {
+		return nil
+	}
+	m := pathRegexp.FindStringSubmatch(input)
+	if len(m) == 0 {
+		return fmt.Errorf("paths can only contain the characters a-zA-Z0-9.-_/")
+	}
+	return nil
+}
+
+func validateStorageConfig(in *manifest.Storage) error {
+	if in == nil {
+		return nil
+	}
+	return validateVolumes(in.Volumes)
+}
+
+func validateVolumes(in map[string]manifest.Volume) error {
+	for name, v := range in {
+		if err := validateVolume(name, v); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func validateVolume(name string, in manifest.Volume) error {
+	if err := validateMountPointConfig(in); err != nil {
+		return fmt.Errorf("validate container configuration for volume %s: %w", name, err)
+	}
+	if err := validateEFSConfig(in); err != nil {
+		return fmt.Errorf("validate EFS configuration for volume %s: %w", name, err)
+	}
+	return nil
+}
+
+func validateMountPointConfig(in manifest.Volume) error {
+	// containerPath must be specified.
+	path := aws.StringValue(in.ContainerPath)
+	if path == "" {
+		return errNoContainerPath
+	}
+	if err := validateContainerPath(path); err != nil {
+		return fmt.Errorf("validate container path %s: %w", path, err)
+	}
+	return nil
+}
+
+func validateSidecarMountPoints(in []manifest.SidecarMountPoint) error {
+	if in == nil {
+		return nil
+	}
+	for _, mp := range in {
+		if aws.StringValue(mp.ContainerPath) == "" {
+			return errNoContainerPath
+		}
+		if aws.StringValue(mp.SourceVolume) == "" {
+			return errNoSourceVolume
+		}
+	}
+	return nil
+}
+
+func validateEFSConfig(in manifest.Volume) error {
+	// EFS is implicitly disabled.
+	if in.EFS == nil {
+		return nil
+	}
+	// This should never happen but error when EFS is Enabled with a non-empty configuration.
+	if aws.BoolValue(in.EFS.Enabled) && !in.EFS.Advanced.IsEmpty() {
+		return errInvalidEFSConfig
+	}
+
+	// If EFS is disabled explicitly, return nil.
+	if in.EFS.Enabled != nil && !aws.BoolValue(in.EFS.Enabled) {
+		return nil
+	}
+
+	// UID and GID are mutually exclusive with any other fields.
+	if !in.EFS.Advanced.EmptyBYOConfig() && !in.EFS.Advanced.EmptyUIDConfig() {
+		return errUIDWithNonManagedFS
+	}
+
+	// Check that required fields for BYO EFS are satisfied.
+	if !in.EFS.Advanced.EmptyBYOConfig() && !in.EFS.Advanced.IsEmpty() {
+		if aws.StringValue(in.EFS.Advanced.FileSystemID) == "" {
+			return errNoFSID
+		}
+	}
+
+	if err := validateRootDirPath(aws.StringValue(in.EFS.Advanced.RootDirectory)); err != nil {
+		return err
+	}
+
+	if err := validateAuthConfig(in.EFS.Advanced); err != nil {
+		return err
+	}
+
+	if err := validateUIDGID(in.EFS.Advanced.UID, in.EFS.Advanced.GID); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func validateAuthConfig(in manifest.EFSVolumeConfiguration) error {
+	if in.AuthConfig == nil {
+		return nil
+	}
+	rd := aws.StringValue(in.RootDirectory)
+	if !(rd == "" || rd == "/") && in.AuthConfig.AccessPointID != nil {
+		return errAccessPointWithRootDirectory
+	}
+
+	if in.AuthConfig.AccessPointID != nil && !aws.BoolValue(in.AuthConfig.IAM) {
+		return errAccessPointWithoutIAM
+	}
+
+	return nil
+}
+
+func validateUIDGID(uid, gid *uint32) error {
+	if uid == nil && gid == nil {
+		return nil
+	}
+	if uid != nil && gid == nil {
+		return errInvalidUIDGIDConfig
+	}
+	if uid == nil && gid != nil {
+		return errInvalidUIDGIDConfig
+	}
+	// Check for root UID.
+	if aws.Uint32Value(uid) == 0 {
+		return errReservedUID
+	}
+	return nil
+}
+
+func validateRootDirPath(input string) error {
+	return validatePath(input, maxEFSPathLength)
+}
+
+func validateContainerPath(input string) error {
+	return validatePath(input, maxDockerContainerPathLength)
+}
